Report index phase counters when a run gives up, and count deletes
Three fixes from review:

A recorder was dropped without reporting when a build or update returned
early, so the metrics were missing exactly the runs that failed. All three
now flush on the way out.

Deletes were never counted as converted, so an update that mostly deleted
looked like it was dropping documents. Building a trash marker is now timed
and counted, and a delete that needs no marker counts too.

Indexed bytes were recorded even when the batch failed to commit. The time
still counts, since it was spent, but the bytes only count once they land.
